+++ packages/core/utils/src/index.ts
@@ -1,7 +1,7 @@
+export * from './date';
 export * from './merge';
 export * from './mixin';
 export * from './mixin/AsyncEmitter';
 export * from './registry';
 export * from './requireModule';
 export * from './uid';
-
